interpolate update and render gap

+++ game_project_perditio/per_vector.h
@@ -9,7 +9,7 @@ struct PERVec3 {
 	PERVec3(double x, double y, double z) : x(x), y(y), z(z) {}
 	PERVec3(const PERVec3& copy) : x(copy.x), y(copy.y), z(copy.z) {}
 
-	PERVec3 operator*(int rhs) {
+	PERVec3 operator*(double rhs) {
 		PERVec3 res;
 		res.x = this->x * rhs;
 		res.y = this->y * rhs;
@@ -26,7 +26,7 @@ struct PERVec2 {
 	PERVec2(double x, double y) : x(x), y(y) {}
 	PERVec2(const PERVec2& copy) : x(copy.x), y(copy.y) {}
 
-	PERVec2 operator*(int rhs) {
+	PERVec2 operator*(double rhs) {
 		PERVec2 res;
 		res.x = this->x * rhs;
 		res.y = this->y * rhs;
